refactor: restructure todo components and improve naming consistency
- Extract TaskList and AddTaskButton into separate components
- Rename TodoList to TodoContainer to better reflect its role as parent component
- Standardize prop names for toggleTask and editTask across components
- Improve component organization and state management
- Move component-specific interfaces to their respective files, keeping only Task interface in types.ts
