--- src/components/ToDoList.tsx
@@ -1,158 +1,0 @@
-import { useState } from "react";
-import {
-  DragDropContext,
-  Droppable,
-  Draggable,
-  DropResult,
-} from "@hello-pangea/dnd";
-import plusCircle from "../assets/icons/plus-circle.svg";
-import TaskForm from "./TaskForm";
-import TaskItem from "./TaskItem";
-import ToDoHeader from "./ToDoHeader";
-import { Task, TaskListProps } from "../types/types";
-import { useTaskManager } from "../hooks/useTaskManager";
-
-export default function ToDoList() {
-  const {
-    tasks,
-    addTask,
-    updateTask,
-    toggleTask,
-    deleteTask,
-    reorderTasks,
-    sortActiveTasksFirst,
-    sortFinishedTasksFirst,
-    deleteFinishedTasks,
-    deleteAllTasks,
-  } = useTaskManager();
-  const [showTaskForm, setShowTaskForm] = useState(false);
-  const [taskToEdit, setTaskToEdit] = useState<Task | null>(null);
-
-  const handleFormClose = () => {
-    setShowTaskForm(false);
-    setTaskToEdit(null);
-  };
-
-  const TaskList = ({ tasks, onToggle, onEdit, onDelete }: TaskListProps) => {
-    const onDragEnd = (result: DropResult) => {
-      if (!result.destination) return;
-      reorderTasks(result.source.index, result.destination.index);
-    };
-
-    return (
-      <DragDropContext onDragEnd={onDragEnd}>
-        <Droppable droppableId="droppable">
-          {(provided) => (
-            <div
-              className="flex w-full max-w-md flex-col"
-              ref={provided.innerRef}
-              {...provided.droppableProps}
-            >
-              {tasks.map((task, index) => (
-                <Draggable key={task.id} draggableId={task.id} index={index}>
-                  {(provided) => (
-                    <div
-                      className="mb-2"
-                      ref={provided.innerRef}
-                      {...provided.draggableProps}
-                      {...provided.dragHandleProps}
-                    >
-                      {taskToEdit?.id === task.id ? (
-                        <TaskForm
-                          key={task.id}
-                          onClose={handleFormClose}
-                          onTaskSubmit={updateTask}
-                          onTaskDelete={deleteTask}
-                          initialTask={taskToEdit}
-                          isEditing={true}
-                        />
-                      ) : (
-                        <div
-                          role="listitem"
-                          key={task.id}
-                          draggable={true}
-                          className="cursor-grab [&.is-dragging]:cursor-grabbing"
-                        >
-                          <TaskItem
-                            key={task.id}
-                            task={task}
-                            onToggle={onToggle}
-                            onEdit={onEdit}
-                            onDelete={onDelete}
-                          />
-                        </div>
-                      )}
-                    </div>
-                  )}
-                </Draggable>
-              ))}
-              {/* This placeholder maintains the space for the dragged item */}
-              {provided.placeholder}
-            </div>
-          )}
-        </Droppable>
-      </DragDropContext>
-    );
-  };
-  interface AddTaskButtonProps {
-    onClick: () => void;
-  }
-
-  const AddTaskButton = ({ onClick }: AddTaskButtonProps) => (
-    <button
-      className="flex w-full cursor-pointer items-center justify-center gap-2 rounded-lg border-3 border-dashed border-gray-100 bg-[#23cc97c6] px-2 py-4 text-gray-100 shadow-xl hover:border-white hover:bg-[#23cc97] hover:text-white"
-      id="addTask"
-      aria-label="add new task"
-      onClick={onClick}
-    >
-      <span>
-        <img src={plusCircle} className="w-6" alt="add" />
-      </span>
-      <span className="text-2xl font-bold">Add Task</span>
-    </button>
-  );
-
-  return (
-    <div
-      className="flex w-full max-w-md flex-col gap-6 p-4"
-      id="toDoListContainer"
-    >
-      <ToDoHeader
-        sortActiveTasksFirst={sortActiveTasksFirst}
-        sortFinishedTasksFirst={sortFinishedTasksFirst}
-        deleteFinishedTasks={deleteFinishedTasks}
-        deleteAllTasks={deleteAllTasks}
-      />
-
-      {showTaskForm && !taskToEdit && (
-        <TaskForm
-          onClose={handleFormClose}
-          onTaskSubmit={addTask}
-          initialTask={null}
-          isEditing={false}
-        />
-      )}
-
-      {!showTaskForm && (
-        <AddTaskButton
-          onClick={() => {
-            if (taskToEdit) {
-              setTaskToEdit(null);
-            }
-            setShowTaskForm(true);
-          }}
-        />
-      )}
-
-      <TaskList
-        onDelete={deleteTask}
-        tasks={tasks}
-        onToggle={toggleTask}
-        onEdit={(task) => {
-          setShowTaskForm(false);
-          setTaskToEdit(task);
-        }}
-      />
-    </div>
-  );
-}
