+++ src/types/types.ts
@@ -4,17 +4,3 @@ export interface Task {
   taskNotes: string;
   finished: boolean;
 }
-
-export interface TaskActionProps {
-  onToggle: (taskId: string) => void;
-  onEdit: (task: Task) => void;
-  onDelete: (taskId: string) => void;
-}
-
-export interface TaskListProps extends TaskActionProps {
-  tasks: Task[];
-}
-
-export interface TaskItemProps extends TaskActionProps {
-  task: Task;
-}
